feat: take data of nearest monitor to the ward
The original code take the data of the monitors in two steps: for the ward with a monitor, just take the data of the monitor; for the wards without a monitor, take the data of the nearest monitor. However, if centroid is regarded at the represence of the ward (as for latitude and longitude information in i-Tree), than it might be more reasonable to take the data of the nearest monitor as the data for a ward. Besides, there are also some wards have more than one monitor, the new feat can simplify the data processing for those data.